pyproject.toml
[tool.poetry]
name = "fame-data-quality-assessment"
version = "0.1.0"
description = ""
authors = ["George Fatouros <gfatouros@innov-acts.com>"]
readme = "README.md"
packages = [{include = "fame_data_quality_assessment"}]

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.31.1"
pandas = "^2.2.1"
openpyxl = "^3.1.2"
sqlalchemy = "^2.0.27"
pyyaml = "^6.0.1"
plotly = "^5.19.0"
pygwalker = "^0.4.6"
seaborn = "^0.13.2"
matplotlib = "^3.8.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
